requirements.txt
astroid==1.4.5
astrolabe==0.4.0
atomic==0.7.2
cffi==1.8.3
click==6.6
colorama==0.3.7
Django==1.8
django-allauth==0.25.2
django-audit-log==0.7.0
django-cors-headers==1.1.0
django-crequest==2014.9.19
django-crontab==0.7.1
django-debug-toolbar==1.4
django-extensions==1.6.7
django-extlog==0.8
django-filter==0.15.3
django-grappelli==2.8.2
django-logentry-admin==1.0.2
django-redis==4.4.4
django-rest-auth==0.7.0
django-rest-swagger==0.3.7
django-websocket-redis==0.4.6
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
future==0.15.2
gevent==1.1.2
git-pylint-commit-hook==2.2.1
gitdb==0.6.4
GitPython==2.0.8
greenlet==0.4.10
gunicorn==19.6.0
Jinja2==2.8
lazy-object-proxy==1.2.2
MarkupSafe==0.23
meld3==1.0.2
metrology==1.2.1
model-mommy==1.3.0
MySQL-python==1.2.5
names==0.3.0
oauthlib==1.1.2
pep8==1.7.0
pycparser==2.17
pydot==1.2.2
pygraphviz==1.3.1
PyJWT==1.4.1
pylint==1.5.5
pyparsing==2.1.8
python-dateutil==2.5.3
python-memcached==1.58
python-mimeparse==1.5.2
python-openid==2.2.5
pytz==2016.4
PyYAML==3.11
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
setproctitle==1.1.10
simplejson==3.8.2
six==1.10.0
smmap==0.9.0
sqlparse==0.1.19
structlog==16.1.0
supervisor==3.3.1
wrapt==1.10.8
XlsxWriter==0.8.9
